Change averageRating from int to decimal in restaurants table

+++ database-mysql/schema.sql
@@ -8,7 +8,7 @@ CREATE TABLE restaurants (
   id int NOT NULL AUTO_INCREMENT,
   name varchar(100) NOT NULL,
   reviewCount int NOT NULL,
-  averageRating smallint NOT NULL,
+  averageRating decimal(2,1) NOT NULL,
   priceRange smallint NOT NULL,
   PRIMARY KEY (id)
 );
